requirements.txt
ninja==1.11.1.4
cmake==4.0.3
scikit-learn==1.5.2
httpx==0.27.0
Jinja2==3.1.4
numpy==1.26.4
openai==1.26.0
pandas==2.2.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
tiktoken==0.6.0
urllib3==2.2.1
python-dotenv==1.1.1